requirements.txt
attrs==19.3.0
freezegun==0.3.15
isort==4.3.21
marshmallow==3.5.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
RPi.GPIO==0.7.0
six==1.14.0
wcwidth==0.1.8
